refs #PMP2-348 : New path connection management
Path management logic has been changed in a way such that each path does
not need anymore to know the global connection status and the existance
of the other paths.

In addition, the path A is no longer labeled the "active" path. Any path
can be "active" provided it is able to send messages and no other path
is "active".

Signed-off-by: FERSULA Jeremy <jeremy.fersula@external.equans.com>